Move String+Helper to WordPressShareD
